Split main.go into focused files
Split 1000-line main.go into separate files by responsibility:
- server.go: Server lifecycle (NewServer, Start, Stop)
- client.go: Client connection handling
- room.go: Room management and broadcasting
- commands.go: Command parsing and handling
- dispatch.go: Message routing loop
- db.go: Repo struct, pool init
- store.go: All database query methods
- models.go: Structs (Client, Server, Room, Message)
- crypto.go: Argon2 hashing and verification
