+++ main.go
@@ -1,1008 +1,16 @@
 package main
 
 import (
-	"bufio"
-	argon2stuff "darkwave/passwordHandling"
-	"fmt"
 	"log"
-	"net"
 	"os"
-	"strings"
-	"sync"
-	"time"
 
 	"context"
 
 	// PostgreSQL driver
-	"github.com/jackc/pgx/v5/pgxpool"
+
 	"github.com/joho/godotenv"
-
 	// needed for password hashing
-	"crypto/rand"
-	"crypto/subtle"
-	"encoding/base64"
-	"errors"
-	"strconv"
-
-	"golang.org/x/crypto/argon2"
 )
-
-type Argon2Params struct {
-	Memory      uint32 //KiB
-	Iterations  uint32
-	Parallelism uint8
-	SaltLength  uint32
-	KeyLength   uint32
-}
-
-var param = &Argon2Params{
-	Memory:      64 * 1024, // MiB
-	Iterations:  3,
-	Parallelism: 2,
-	SaltLength:  16,
-	KeyLength:   32,
-}
-
-type Client struct {
-	UserID     int64
-	conn       net.Conn
-	out        chan string
-	nick       string
-	password   string
-	activeRoom *Room
-	friends    []Friend
-	friendReqs []FriendRequest
-	wmu        sync.Mutex
-}
-
-type FriendRequest struct {
-	ID          int64
-	fromNick    *Client
-	toNick      *Client
-	fromID      int64  // new
-	toID        int64  // new
-	status      string // pending, accepted, declined
-	createdAt   time.Time
-	respondedAt *time.Time
-}
-
-type Friend struct {
-	friendID  int64   // new
-	userID    int64   // new
-	friend    *Client // old
-	createdAt time.Time
-}
-
-type Room struct {
-	ID      int64
-	Name    string
-	Inbox   chan Message
-	clients map[*Client]struct{}
-	mu      sync.RWMutex
-}
-
-type Message struct {
-	ID      int64
-	room    *Room
-	from    *Client
-	payload []byte
-}
-
-type DBMessage struct {
-	ID     int64
-	RoomID int64
-	FromID int64
-	Body   string
-	SentAt time.Time
-}
-
-type MessageWithSender struct {
-	Body   string
-	SentAt time.Time
-	Nick   string
-}
-
-type Repo struct {
-	pool *pgxpool.Pool
-}
-
-type Server struct {
-	listenAddr string
-	ln         net.Listener
-	quitch     chan struct{}
-
-	// rooms registry
-	mu            sync.RWMutex
-	rooms         map[string]*Room
-	clientsByUser map[int64]*Client
-	repo          *Repo
-
-	// global inbox (roomed messages)
-	msgch chan Message
-}
-
-func NewRepo(p *pgxpool.Pool) *Repo {
-	return &Repo{pool: p}
-}
-
-func NewServer(addr string, repo *Repo) *Server {
-	s := &Server{
-		listenAddr:    addr,
-		quitch:        make(chan struct{}),
-		msgch:         make(chan Message, 1024),
-		rooms:         make(map[string]*Room),
-		clientsByUser: make(map[int64]*Client),
-		repo:          repo,
-	}
-	// create #main with a bigger buffer (absorbs bursts)
-	s.rooms["#main"] = &Room{
-		ID:      1,
-		Name:    "#main",
-		Inbox:   make(chan Message, 1024),
-		clients: make(map[*Client]struct{}),
-	}
-	return s
-}
-
-func (s *Server) Start(ctx context.Context) error {
-	ln, err := net.Listen("tcp", s.listenAddr)
-	if err != nil {
-		return err
-	}
-	defer ln.Close()
-	s.ln = ln
-	fmt.Println("listening on", ln.Addr().String())
-
-	// one broadcaster per room
-	for _, r := range s.rooms {
-		go s.roomBroadcastLoop(r)
-	}
-	// global dispatcher: takes Message (with room) and pushes into that room’s Inbox
-	go s.dispatchLoop()
-
-	// accept clients
-	go s.acceptLoop(ctx)
-
-	<-ctx.Done()
-	return nil
-}
-
-func (s *Server) Stop() {
-	select {
-	case s.quitch <- struct{}{}:
-	default:
-	}
-	if s.ln != nil {
-		_ = s.ln.Close()
-	}
-}
-
-func (s *Server) acceptLoop(ctx context.Context) {
-	for {
-		conn, err := s.ln.Accept()
-		if err != nil {
-			select {
-			case <-ctx.Done():
-				return // clean shutdown
-			default:
-				fmt.Println("accept:", err)
-				close(s.quitch) // unexpected error, signal shutdown
-				return
-			}
-		}
-
-		host, port, _ := net.SplitHostPort(conn.RemoteAddr().String())
-		_ = host
-		c := &Client{
-			conn: conn,
-			out:  make(chan string, 32),
-			nick: "guest-" + port,
-		}
-
-		go s.writeLoop(c)
-		go s.readLoop(c)
-	}
-}
-
-func (s *Server) getOrCreateRoom(name string) *Room {
-	if name == "" {
-		name = "#main"
-	}
-	if !strings.HasPrefix(name, "#") && !strings.HasPrefix(name, "dm:") {
-		name = "#" + name
-	}
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if r, ok := s.rooms[name]; ok {
-		return r
-	}
-	r := &Room{
-		Name:    name,
-		Inbox:   make(chan Message, 256),
-		clients: make(map[*Client]struct{}),
-	}
-	s.rooms[name] = r
-	go s.roomBroadcastLoop(r)
-	return r
-}
-
-func (s *Server) dispatchLoop() {
-	for m := range s.msgch {
-
-		// go func(msg Message) {       <--- - if we wanted to save asynchronously (not blocking the dispatch loop)
-		if err := s.repo.SaveMessage(m); err != nil {
-			fmt.Println("failed to save message:", err)
-		}
-		// }(m)
-
-		// forward to the room’s inbox (copy is already done in readLoop)
-		m.room.Inbox <- m
-
-		// server-side log
-		from := m.from.nick
-		fmt.Printf("DISPATCH -> %s\n", m.room.Name)
-		fmt.Printf("[%s] (%s) %s", m.room.Name, from, string(m.payload))
-	}
-}
-
-func (s *Server) roomBroadcastLoop(r *Room) {
-	r.mu.RLock()
-	fmt.Printf("BROADCAST %s to %d clients\n", r.Name, len(r.clients))
-	r.mu.RUnlock()
-	for msg := range r.Inbox {
-		r.mu.RLock()
-		clients := make([]*Client, 0, len(r.clients))
-		for c := range r.clients {
-			clients = append(clients, c)
-		}
-		r.mu.RUnlock() // release lock before sending
-		for _, c := range clients {
-			line := fmt.Sprintf("[%s] (%s) %s", msg.room.Name, msg.from.nick, strings.TrimSpace(string(msg.payload)))
-			s.sendLine(c, "%s", line)
-		}
-	}
-}
-
-func (s *Server) joinRoom(c *Client, r *Room) {
-	if c.activeRoom != nil {
-		old := c.activeRoom
-		old.mu.Lock()
-		delete(old.clients, c)
-		old.mu.Unlock()
-	}
-	r.mu.Lock()
-	r.clients[c] = struct{}{}
-	r.mu.Unlock()
-	c.activeRoom = r
-
-	s.sendLine(c, "** joined %s", r.Name) // newline guaranteed
-}
-
-func (r *Repo) listRooms() ([]string, error) {
-	rows, err := r.pool.Query(context.Background(), `
-		select name from rooms order by name
-	`)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	var names []string
-	for rows.Next() {
-		var name string
-		if err := rows.Scan(&name); err != nil {
-			return nil, err
-		}
-		names = append(names, name)
-	}
-	return names, nil
-}
-
-func cleanInput(s string) string {
-	s = strings.TrimRight(s, "\r\n")
-	s = strings.TrimSpace(s)
-	// strip BOM and common zero-width chars at the start
-	s = strings.TrimPrefix(s, "\uFEFF") // BOM
-	s = strings.TrimPrefix(s, "\u200B") // zero-width space
-	s = strings.TrimPrefix(s, "\u200C") // ZWNJ
-	s = strings.TrimPrefix(s, "\u200D") // ZWJ
-	s = strings.TrimPrefix(s, "\u2060") // word joiner
-	return s
-}
-
-func (s *Server) readLoop(c *Client) {
-	defer func() {
-		s.unregisterClientSingle(c)
-		if c.activeRoom != nil {
-			c.activeRoom.mu.Lock()
-			delete(c.activeRoom.clients, c)
-			c.activeRoom.mu.Unlock()
-		}
-		close(c.out)
-		_ = c.conn.Close()
-		fmt.Println("client disconnected:", c.nick)
-	}()
-
-	r := bufio.NewReader(c.conn)
-	for {
-		line, err := r.ReadString('\n')
-		if err != nil {
-			return
-		}
-
-		text := cleanInput(line)
-		if text == "" {
-			continue
-		}
-
-		if strings.HasPrefix(text, "/") {
-			fmt.Println("DISPATCH CMD:", strconv.Quote(text)) // add: proves we hit command path
-			s.handleCommand(c, text)
-			continue
-		}
-
-		if c.UserID == 0 {
-			s.sendLine(c, "** set your nick first with /nick <name>")
-			continue
-		}
-		if c.activeRoom == nil {
-			s.sendLine(c, "** join a room first with /join <room>")
-			continue
-		}
-
-		out := append([]byte(nil), (text + "\n")...)
-
-		s.msgch <- Message{room: c.activeRoom, from: c, payload: out}
-	}
-}
-
-func (s *Server) writeLoop(c *Client) {
-	for line := range c.out {
-		if _, err := c.conn.Write([]byte(line + "\n")); err != nil {
-			return
-		}
-	}
-}
-
-func (s *Server) sendLine(c *Client, format string, args ...any) error {
-	c.wmu.Lock()
-	defer c.wmu.Unlock()
-
-	// Optional: detect closed conn early
-	// _ = c.conn.SetWriteDeadline(time.Now().Add(2 * time.Second))
-
-	n, err := fmt.Fprintf(c.conn, format+"\n", args...) // newline required
-	if err != nil {
-		fmt.Println("sendLine write error:", err)
-		return err
-	}
-	fmt.Printf("sendLine wrote %d bytes: %q\n", n, fmt.Sprintf(format, args...))
-	return nil
-}
-
-func (s *Server) registerClientSingle(c *Client) {
-	s.mu.Lock()
-	if old := s.clientsByUser[c.UserID]; old != nil && old != c {
-		_ = s.sendLine(old, "** you were logged out (new login elsewhere)")
-		_ = old.conn.Close() // triggers its cleanup defer
-	}
-	s.clientsByUser[c.UserID] = c
-	s.mu.Unlock()
-}
-
-func (s *Server) unregisterClientSingle(c *Client) {
-	s.mu.Lock()
-	if s.clientsByUser[c.UserID] == c {
-		delete(s.clientsByUser, c.UserID)
-	}
-	s.mu.Unlock()
-}
-
-func (s *Server) onlineByUserID(userID int64) (*Client, bool) {
-	s.mu.RLock()
-	cc, ok := s.clientsByUser[userID]
-	s.mu.RUnlock()
-	return cc, ok
-}
-
-// create-or-get client by nick
-func (r *Repo) UpsertClientByNick(ctx context.Context, nick string) (id int64, retNick string, err error) {
-	err = r.pool.QueryRow(ctx, `
-        insert into clients(nick) values ($1)
-        on conflict(nick) do update set nick = excluded.nick
-        returning id, nick
-    `, nick).Scan(&id, &retNick)
-	return id, retNick, err
-}
-
-// create-or-get room by name
-func (r *Repo) UpsertRoomByName(ctx context.Context, name string) (id int64, retName string, err error) {
-	err = r.pool.QueryRow(ctx, `
-        insert into rooms(name) values ($1)
-        on conflict(name) do update set name = excluded.name
-        returning id, name
-    `, name).Scan(&id, &retName)
-	return id, retName, err
-}
-
-func (r *Repo) declineFriendRequest(ctx context.Context, fromID, toID int64) error {
-	_, err := r.pool.Exec(ctx, `
-		update friend_requests
-		set status = 'declined', responded_at = now()
-		where from_id = $1 and to_id = $2 and status = 'pending'
-	`, fromID, toID)
-	return err
-}
-
-func (r *Repo) addFriend(ctx context.Context, fromID, toID int64) error {
-	_, err := r.pool.Exec(ctx, `
-		insert into friends(user_id, friend_id, created_at)
-		values ($1, $2, now())
-	`, fromID, toID)
-	return err
-}
-
-func (r *Repo) acceptFriendRequest(ctx context.Context, fromID, toID int64) error {
-	_, err := r.pool.Exec(ctx, `
-		update friend_requests
-		set status = 'accepted', responded_at = now()
-		where from_id = $1 and to_id = $2 and status = 'pending'
-	`, fromID, toID)
-
-	return err
-}
-
-func (r *Repo) getFriendsByUserID(ctx context.Context, userID int64) ([]Friend, error) {
-	rows, err := r.pool.Query(ctx, `
-		select friend_id, created_at from friends where user_id = $1
-	`, userID)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var friends []Friend
-	for rows.Next() {
-		var f Friend
-		if err := rows.Scan(&f.friendID, &f.createdAt); err != nil {
-			return nil, err
-		}
-		friends = append(friends, f)
-	}
-	return friends, nil
-}
-
-func (r *Repo) getFriendRequestsByUserID(ctx context.Context, userID int64) ([]FriendRequest, error) {
-	rows, err := r.pool.Query(ctx, `
-		select id, from_id, to_id, status, created_at, responded_at
-		from friend_requests
-		where to_id = $1 and status = 'pending'
-	`, userID)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-	var requests []FriendRequest
-	for rows.Next() {
-		var fr FriendRequest
-		if err := rows.Scan(&fr.ID, &fr.fromID, &fr.toID, &fr.status, &fr.createdAt, &fr.respondedAt); err != nil {
-			return nil, err
-		}
-		requests = append(requests, fr)
-	}
-	return requests, nil
-}
-
-func (r *Repo) sendFriendRequest(fromID, toID int64) error {
-	_, err := r.pool.Exec(context.Background(), `
-		insert into friend_requests(from_id, to_id, status, created_at)
-		values ($1, $2, 'pending', now())
-	`, fromID, toID)
-	return err
-}
-
-func (r *Repo) findClientIDByNick(nick string) (int64, error) {
-	var id int64
-	err := r.pool.QueryRow(context.Background(), `
-		select id from clients where nick = $1
-	`, nick).Scan(&id)
-	if err != nil {
-		return 0, err
-	}
-	return id, nil
-}
-
-func (r *Repo) checkClientHasPassword(ctx context.Context, id int64) (bool, error) {
-	var storedHash string
-	err := r.pool.QueryRow(ctx, `
-		select password from clients where id = $1
-	`, id).Scan(&storedHash)
-	if err != nil {
-		return false, err
-	}
-	return storedHash != "", nil
-}
-
-func (r *Repo) getClientByID(ctx context.Context, id int64) (*Client, error) {
-	var nick string
-	err := r.pool.QueryRow(ctx, `
-		select nick from clients where id = $1
-	`, id).Scan(&nick)
-	if err != nil {
-		return nil, err
-	}
-	return &Client{UserID: id, nick: nick}, nil
-}
-
-func HashPassword(password string, p *Argon2Params) (encodedHash string, err error) {
-	salt := make([]byte, p.SaltLength)
-	if _, err := rand.Read(salt); err != nil {
-		return "", err
-	}
-
-	hash := argon2.IDKey([]byte(password), salt, p.Iterations, p.Memory, p.Parallelism, p.KeyLength)
-
-	b64Salt := base64.RawStdEncoding.EncodeToString(salt)
-	b64Hash := base64.RawStdEncoding.EncodeToString(hash)
-
-	encodedHash = fmt.Sprintf("$argon2id$v=19$m=%d,t=%d,p=%d$%s$%s", p.Memory, p.Iterations, p.Parallelism, b64Salt, b64Hash)
-	return encodedHash, nil
-}
-
-func VerifyPassword(password, encodedHash string) (bool, error) {
-	if password == "" {
-		return false, errors.New("empty password")
-	}
-	if strings.TrimSpace(encodedHash) == "" {
-		return false, errors.New("empty stored hash")
-	}
-
-	p, salt, expectedHash, err := DecodeHash(encodedHash)
-	if err != nil {
-		return false, err
-	}
-	if len(salt) < 8 {
-		return false, errors.New("bad salt")
-	}
-	if len(expectedHash) == 0 {
-		return false, errors.New("bad hash")
-	}
-
-	// Use the stored hash length as the key length to derive for verification
-	keyLen := uint32(len(expectedHash))
-	key := argon2.IDKey([]byte(password), salt, p.Iterations, p.Memory, p.Parallelism, keyLen)
-
-	if subtle.ConstantTimeCompare(key, expectedHash) == 1 {
-		return true, nil
-	}
-	return false, nil
-}
-
-func DecodeHash(encodedHash string) (Argon2Params, []byte, []byte, error) {
-	// Expected: $argon2id$v=19$m=65536,t=3,p=2$<saltB64>$<hashB64>
-	parts := strings.Split(strings.TrimSpace(encodedHash), "$")
-	if len(parts) != 6 || parts[1] != "argon2id" {
-		return Argon2Params{}, nil, nil, errors.New("invalid encoded hash format")
-	}
-	if parts[2] != "v=19" {
-		return Argon2Params{}, nil, nil, errors.New("unsupported argon2 version")
-	}
-
-	// Parse m=...,t=...,p=...
-	var p Argon2Params
-	items := strings.Split(parts[3], ",")
-	if len(items) != 3 {
-		return Argon2Params{}, nil, nil, errors.New("invalid argon2 parameters")
-	}
-	var err error
-	p.Memory, err = parseUint32KV(items[0], "m")
-	if err != nil {
-		return Argon2Params{}, nil, nil, err
-	}
-	p.Iterations, err = parseUint32KV(items[1], "t")
-	if err != nil {
-		return Argon2Params{}, nil, nil, err
-	}
-	par, err := parseUint32KV(items[2], "p")
-	if err != nil {
-		return Argon2Params{}, nil, nil, err
-	}
-	if par > 255 {
-		return Argon2Params{}, nil, nil, errors.New("invalid parallelism value")
-	}
-	p.Parallelism = uint8(par)
-
-	salt, err := base64.RawStdEncoding.DecodeString(parts[4])
-	if err != nil {
-		return Argon2Params{}, nil, nil, errors.New("invalid salt b64")
-	}
-
-	hash, err := base64.RawStdEncoding.DecodeString(parts[5])
-	if err != nil {
-		return Argon2Params{}, nil, nil, errors.New("invalid hash b64")
-	}
-
-	// Set key length from stored hash length
-	p.KeyLength = uint32(len(hash))
-
-	return p, salt, hash, nil
-}
-
-func parseUint32KV(s, key string) (uint32, error) {
-	kv := strings.SplitN(s, "=", 2)
-	if len(kv) != 2 || kv[0] != key {
-		return 0, errors.New("invalid format for " + key)
-	}
-	n, err := strconv.ParseUint(kv[1], 10, 32)
-	if err != nil {
-		return 0, err
-	}
-	return uint32(n), nil
-}
-
-func (s *Server) readLoopPassword(c *Client) (string, error) {
-
-	r := bufio.NewReader(c.conn)
-	line, err := r.ReadString('\n')
-	if err != nil {
-		return "** error: %s", err
-	}
-
-	passwordPlainText := cleanInput(line)
-	for i := 0; 5 > len(passwordPlainText); i++ {
-		// Simulate password input processing
-		s.sendLine(c, "** password should be at least 5 characters")
-		passwordPlainText, err = r.ReadString('\n')
-		if err != nil {
-			return "** error: %s", err
-		}
-
-	}
-
-	return passwordPlainText, nil
-}
-
-func (r *Repo) SavePassword(password string, id int64) error {
-	_, err := r.pool.Exec(context.Background(), `
-		UPDATE clients SET password = $1 WHERE id = $2
-	`, password, id)
-	return err
-}
-
-func (r *Repo) GetPasswordById(id int64) (string, error) {
-	var password string
-	err := r.pool.QueryRow(context.Background(), `
-		select password from clients where id = $1
-	`, id).Scan(&password)
-	if err != nil {
-		return "", err
-	}
-	return password, nil
-}
-
-func (r *Repo) SaveMessage(msg Message) error {
-	_, err := r.pool.Exec(context.Background(), `
-		INSERT INTO messages (room_id, from_id, body)
-		VALUES ($1, $2, $3)
-	`,
-		msg.room.ID,
-		msg.from.UserID,
-		string(msg.payload),
-	)
-	return err
-}
-
-func (r *Repo) GetRecentMessages(roomID int64) ([]DBMessage, error) {
-	rows, err := r.pool.Query(context.Background(), `
-		SELECT id, room_id, from_id, body, sent_at
-		FROM messages
-		WHERE room_id = $1
-		ORDER BY sent_at DESC
-		LIMIT 50
-	`, roomID)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var msgs []DBMessage
-	for rows.Next() {
-		var m DBMessage
-		if err := rows.Scan(&m.ID, &m.RoomID, &m.FromID, &m.Body, &m.SentAt); err != nil {
-			return nil, err
-		}
-		msgs = append(msgs, m)
-	}
-	return msgs, nil
-}
-
-func (r *Repo) GetRecentMessagesWithUsers(roomID int64) ([]MessageWithSender, error) {
-	rows, err := r.pool.Query(context.Background(), `
-		SELECT m.body, m.sent_at, c.nick
-		FROM messages m
-		JOIN clients c ON m.from_id = c.id
-		WHERE m.room_id = $1
-		ORDER BY m.sent_at DESC
-		LIMIT 50
-	`, roomID)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	var msgs []MessageWithSender
-	for rows.Next() {
-		var m MessageWithSender
-		if err := rows.Scan(&m.Body, &m.SentAt, &m.Nick); err != nil {
-			return nil, err
-		}
-		msgs = append(msgs, m)
-	}
-	return msgs, nil
-}
-
-const helpText = "" +
-	"** commands **\n" +
-	"  /help | /?            show this help\n" +
-	"  /nick <name>          set your nickname\n" +
-	"  /rooms                list available rooms\n" +
-	"  /join <room>          join or create a room (e.g., /join main)\n" +
-	"  /friends              list your friends\n" +
-	"  /addfriend <nick>     add a friend by nickname\n" +
-	"  /friendreqs           list pending friend requests\n" +
-	"  /acceptfriend <nick>  accept a friend request\n" +
-	"  /declinefriend <nick> decline a friend request\n" +
-	"  /quit                 disconnect\n"
-
-func (s *Server) handleCommand(c *Client, cmd string) {
-	parts := strings.Fields(cmd)
-	if len(parts) == 0 {
-		return
-	}
-
-	name := parts[0]
-
-	fmt.Println("CMD NAME:", name)
-
-	switch name {
-	case "/help", "/?":
-		if err := s.sendLine(c, helpText); err != nil {
-			fmt.Println("sendLine /help:", err)
-		}
-
-	case "/rooms":
-		names, err := s.repo.listRooms()
-		if err != nil {
-			s.sendLine(c, "** error: %v", err)
-			return
-		}
-		var out string
-		if len(names) == 0 {
-			out = "** rooms: (none)"
-		} else {
-			out = "** rooms: " + strings.Join(names, ", ")
-		}
-		if err := s.sendLine(c, "%s", out); err != nil {
-			fmt.Println("sendLine /rooms:", err)
-		}
-	case "/nick":
-		if len(parts) < 2 {
-			s.sendLine(c, "usage: /nick <name>")
-			return
-		}
-		id, retNick, err := s.repo.UpsertClientByNick(context.Background(), parts[1])
-		if err != nil {
-			s.sendLine(c, "** error setting nick: %v", err)
-			return
-		}
-
-		user, err := s.repo.getClientByID(context.Background(), id)
-		if err != nil {
-			s.sendLine(c, "** error fetching client: %v", err)
-			return
-		}
-
-		//input password
-		passwordPlainText, err := s.readLoopPassword(c)
-		if err != nil {
-			s.sendLine(c, "** error reading password: %s", err)
-			return
-		}
-
-		//check if client has password already
-		userHasPassword, err := s.repo.checkClientHasPassword(context.Background(), id)
-		if err != nil {
-			s.sendLine(c, "** error checking password: %s", err)
-			return
-		}
-
-		//has a password already
-		if !userHasPassword {
-			user.nick = retNick
-			user.UserID = id
-			c.nick = retNick
-			c.UserID = id
-
-			//hashing plain text password
-			password, err := argon2stuff.HashPassword(passwordPlainText, (*argon2stuff.Argon2Params)(param))
-			if err != nil {
-				s.sendLine(c, "** error hashing password: %s", err)
-				return
-			}
-
-			if err := s.repo.SavePassword(password, id); err != nil {
-				s.sendLine(c, "** error saving password: %v", err)
-				return
-			}
-		} else {
-			c.nick = user.nick
-			c.UserID = user.UserID
-			dbPassword, err := s.repo.GetPasswordById(id)
-			if err != nil {
-				s.sendLine(c, "** error fetching stored password: %s", err)
-				return
-			}
-
-			for {
-				match, err := VerifyPassword(passwordPlainText, dbPassword)
-				if err != nil {
-					s.sendLine(c, "** error verifying password: %s", err)
-					return
-				}
-				if match {
-					break
-				}
-				s.sendLine(c, "** please re-enter password for %s:", retNick)
-				passwordPlainText, err = s.readLoopPassword(c)
-				if err != nil {
-					s.sendLine(c, "** error reading password: %s", err)
-					return
-				}
-			}
-		}
-
-		s.sendLine(c, "** welcome, %s!", retNick)
-		s.registerClientSingle(c)
-
-		mainRoom := s.getOrCreateRoom("#main")
-		s.joinRoom(c, mainRoom)
-		_ = s.sendLine(c, "** connected. type /help | /?")
-
-		// load recent messages
-		msgs, err := s.repo.GetRecentMessagesWithUsers(mainRoom.ID)
-		if err != nil {
-			s.sendLine(c, "** error fetching recent messages: %v", err)
-		} else {
-			for _, msg := range msgs {
-				s.sendLine(c, "[%s] %s: %s", msg.SentAt.Format("15:04"), msg.Nick, msg.Body)
-			}
-		}
-
-	case "/join":
-		if len(parts) < 2 {
-			s.sendLine(c, "usage: /join <room>")
-			return
-		}
-		id, retName, err := s.repo.UpsertRoomByName(context.Background(), parts[1])
-		if err != nil {
-			s.sendLine(c, "** error: %v", err)
-			return
-		}
-
-		room := s.getOrCreateRoom(retName)
-		room.ID = id
-		room.Name = retName
-		s.joinRoom(c, room)
-		// load recent messages
-		msgs, err := s.repo.GetRecentMessagesWithUsers(room.ID)
-		if err != nil {
-			s.sendLine(c, "** error fetching recent messages: %v", err)
-		} else {
-			for _, msg := range msgs {
-				s.sendLine(c, "[%s] %s: %s", msg.SentAt.Format("15:04"), msg.Nick, msg.Body)
-			}
-		}
-	case "/friends":
-		friends, err := s.repo.getFriendsByUserID(context.Background(), c.UserID)
-
-		if err != nil {
-			s.sendLine(c, "** error fetching friends: %v", err)
-			return
-		}
-		var out string
-		if len(friends) == 0 {
-			out = "** friends: (none)"
-		} else {
-			var names []string
-			for _, f := range friends {
-				names = append(names, f.friend.nick)
-			}
-			out = "** friends: " + strings.Join(names, ", ")
-		}
-		if err := s.sendLine(c, "%s", out); err != nil {
-			fmt.Println("sendLine /friends:", err)
-		}
-	case "/addfriend":
-		if len(parts) < 2 {
-			s.sendLine(c, "usage: /addfriend <nick>")
-			return
-		}
-		nick := parts[1]
-
-		toID, err := s.repo.findClientIDByNick(nick)
-		if err != nil {
-			s.sendLine(c, "** no user with nick: %s", nick)
-			return
-		}
-		s.repo.sendFriendRequest(c.UserID, toID)
-		s.sendLine(c, "** friend request sent to: %s", nick)
-	case "/friendreqs":
-
-		reqs, err := s.repo.getFriendRequestsByUserID(context.Background(), c.UserID)
-		if err != nil {
-			s.sendLine(c, "** error fetching friend requests: %v", err)
-			return
-		}
-		var out string
-		if len(reqs) == 0 {
-			out = "** friend requests: (none)"
-		} else {
-			var names []string
-			for _, fr := range reqs {
-				names = append(names, fr.fromNick.nick)
-			}
-			out = "** friend requests: " + strings.Join(names, ", ")
-		}
-		if err := s.sendLine(c, "%s", out); err != nil {
-			fmt.Println("sendLine /friendreqs:", err)
-		}
-	case "/acceptfriend":
-
-		if len(parts) < 2 {
-			s.sendLine(c, "usage: /acceptfriend <nick>")
-			return
-		}
-
-		nick := parts[1]
-		fromID, err := s.repo.findClientIDByNick(nick)
-		if err != nil {
-			s.sendLine(c, "** no user with nick: %s", nick)
-			return
-		}
-		if err := s.repo.acceptFriendRequest(context.Background(), fromID, c.UserID); err != nil {
-			s.sendLine(c, "** error accepting friend request from: %s", nick)
-			return
-		}
-		if err := s.repo.addFriend(context.Background(), fromID, c.UserID); err != nil {
-			s.sendLine(c, "** error adding friend: %s", nick)
-			return
-		}
-		s.sendLine(c, "** accepted friend request from: %s", nick)
-	case "/declinefriend":
-
-		if len(parts) < 2 {
-			s.sendLine(c, "usage: /declinefriend <nick>")
-			return
-		}
-
-		nick := parts[1]
-		fromID, err := s.repo.findClientIDByNick(nick)
-		if err != nil {
-			s.sendLine(c, "** no user with nick: %s", nick)
-			return
-		}
-		if err := s.repo.declineFriendRequest(context.Background(), fromID, c.UserID); err != nil {
-			s.sendLine(c, "** error declining friend request from: %s", nick)
-			return
-		}
-		s.sendLine(c, "** declined friend request from: %s", nick)
-	case "/quit":
-		s.sendLine(c, "** bye")
-		_ = c.conn.Close()
-
-	default:
-		s.sendLine(c, "** unknown command: %s (type /help | /?)", name)
-	}
-}
 
 func mustEnv(key string) string {
 	v := os.Getenv(key)
@@ -1021,24 +29,10 @@ func main() {
 		port = ":3000"
 	}
 
-	cfg, err := pgxpool.ParseConfig(dbURL)
-	if err != nil {
-		log.Fatal("parse DB_URL:", err)
-	}
-	pool, err := pgxpool.NewWithConfig(context.Background(), cfg)
-	if err != nil {
-		log.Fatal("connect:", err)
-	}
+	pool := mustInitPool(dbURL)
 	defer pool.Close()
 
-	if err := pool.Ping(context.Background()); err != nil {
-		log.Fatal("db ping failed:", err)
-	}
-
-	fmt.Println("DB connected ✅  listening on", port)
-
 	repo := NewRepo(pool)
-
-	s := NewServer(":3000", repo)
+	s := NewServer(port, repo)
 	log.Fatal(s.Start(context.Background()))
 }
